renamed roller relayer channel show to roller relayer status
