+++ cmd/relayer/relayer.go
@@ -1,8 +1,8 @@
 package relayer
 
 import (
-	"github.com/dymensionxyz/roller/cmd/relayer/channel"
-	start "github.com/dymensionxyz/roller/cmd/relayer/start"
+	"github.com/dymensionxyz/roller/cmd/relayer/start"
+	"github.com/dymensionxyz/roller/cmd/relayer/status"
 	"github.com/spf13/cobra"
 )
 
@@ -12,6 +12,6 @@ func Cmd() *cobra.Command {
 		Short: "Commands for running and managing the RollApp relayer.",
 	}
 	cmd.AddCommand(start.Start())
-	cmd.AddCommand(channel.Cmd())
+	cmd.AddCommand(status.Cmd())
 	return cmd
 }
